requirements/base.txt
Babel==2.12.1
Django==4.2
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
Pillow==9.5.0
PySocks==1.7.1
PyYAML==6.0
Pygments==2.14.0
SQLAlchemy==1.4.47
Sphinx==6.1.3
alabaster==0.7.13
amqp==5.1.1
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
billiard==3.6.4.0
boto3==1.26.107
botocore==1.29.107
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cron-descriptor==1.2.35
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
django-bootstrap-datepicker-plus==5.0.3
django-braces==1.15.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-crispy-forms==2.0
django-datetime-widget==0.9.3
django-environ==0.10.0
django-etc==1.4.0
django-extensions==3.2.1
django-formtools==2.4
django-import-export==3.1.0
django-jquery==3.1.0
django-redis==5.2.0
django-silk==5.0.3
django-siteprefs==1.2.3
django-storages==1.13.2
django-tables2==2.5.3
django-timezone-field==5.0
django-tinymce==3.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djfernet==0.8.1
dnspython==2.3.0
docutils==0.19
drf-yasg==1.21.5
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fsspec==2023.3.0
future==0.18.3
gprof2dot==2022.7.29
h11==0.14.0
html2text==2020.1.16
httplib2==0.22.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.1.0
inflection==0.5.1
itypes==1.2.0
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
numpy==1.24.2
oauth2==1.9.0.post1
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==2.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pyOpenSSL==23.1.1
pyathena==2.23.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
s3transfer==0.6.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
supervisor==4.2.5
tablib==3.4.0
tenacity==8.2.2
tomli==2.0.1
tornado==6.2
trio-websocket==0.10.2
trio==0.22.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
